+++ api/src/main/java/ar/edu/itba/paw/api/controller/SongController.java
@@ -66,7 +66,7 @@ public class SongController extends BaseController {
 
         List<SongDTO> songDTOs = songService.findPaginated(filter, page, size);
         List<SongResource> songResources = songResourceMapper.toResourceList(songDTOs, getBaseUrl());
-        Long totalCount = songService.countAll();
+        Integer totalCount = songService.countAll().intValue();
         
         CollectionResource<SongResource> collection = collectionResourceMapper.createCollection(
                 songResources, totalCount, page, size, getBaseUrl(), ApiUriConstants.SONGS_BASE, ControllerUtils.songsCollectionLinks);
@@ -77,7 +77,7 @@ public class SongController extends BaseController {
     private Response getSongBySubstring(String substring, Integer page, Integer size) {
         List<SongDTO> songDTOs = songService.findByTitleContaining(substring, page, size);
         List<SongResource> songResources = songResourceMapper.toResourceList(songDTOs, getBaseUrl());
-        Long totalCount = songService.countAll();
+        Integer totalCount = songService.countAll().intValue();
         CollectionResource<SongResource> collection = collectionResourceMapper.createCollection(
                 songResources, totalCount, page, size, getBaseUrl(), ApiUriConstants.SONGS_BASE, ControllerUtils.songsCollectionLinks);
         return buildResponse(collection);
@@ -131,7 +131,7 @@ public class SongController extends BaseController {
         List<ReviewDTO> reviews = reviewService.findSongReviewsPaginated(id, page, size, loggedUserId);
 
         List<ReviewResource> reviewResources = reviewResourceMapper.toResourceList(reviews, getBaseUrl());
-        Long totalCount = reviewService.countAll();
+        Integer totalCount = reviewService.countAll().intValue();
                 
         CollectionResource<ReviewResource> collection = collectionResourceMapper.createCollection(
                     reviewResources, totalCount, page, size, getBaseUrl(), ApiUriConstants.SONGS_BASE + ApiUriConstants.SONG_REVIEWS, ControllerUtils.itemReviewsCollectionLinks, id);
